Extract advanceDateToSkipBadDataIfNecessary to shared code
As suggested by @gsnedders in https://github.com/Ecosystem-Infra/wpt-results-analysis/pull/73#discussion_r798878538
